requirements.txt
keyboard==0.13.5
mcpi==1.2.1
RPi.GPIO==0.7.0